# [WIP] perf: early check for .git directory to reduce check time