@JanDeDobbeleer 2025-06-29
@nopeless using go-git is a no go for me as it's not necessarily faster. I'll check how to solve this by having a look tomorrow.