@JanDeDobbeleer 2025-06-30
part of #6567